server/requirements.txt
fastapi
uvicorn[standard]
beancount
pyyaml
bcrypt
python-jose[cryptography]
